Add support for multiple select values in core
