+++ compat/src/index.js
@@ -128,14 +128,6 @@ function createElement(...args) {
 			delete props.defaultValue;
 		}
 
-		if (Array.isArray(props.value) && props.multiple && type==='select') {
-			toChildArray(props.children).forEach((child) => {
-				if (props.value.indexOf(child.props.value)!==-1) {
-					child.props.selected = true;
-				}
-			});
-			delete props.value;
-		}
 		handleElementVNode(vnode, props);
 	}
